# remove network dependencies from explainer