## Checklist

- [ ] Update changelog in CHANGELOG.md.



## Context 

This PR removes all network dependencies from the explainers. The problem with having explainers depend on external resources is increased latency and in turn a worse user experience. With this PR execution times of even only printing the help usage dropped from **0.738** to **0.038** seconds. 



#### Before 

<pre>
$ time awscnfm cl001 ac001 -h
Create a basic Tenant Cluster with all its defaults.

Usage:
  awscnfm cl001 ac001 [flags]

Flags:
  -h, --help   help for ac001
awscnfm cl001 ac001 -h  0.13s user 0.07s system 27% cpu <b>0.738</b> total
</pre>



#### After 

<pre>
$ time awscnfm cl001 ac001 -h
Create a basic Tenant Cluster with all its defaults.

Usage:
  awscnfm cl001 ac001 [flags]

Flags:
  -h, --help   help for ac001
awscnfm cl001 ac001 -h  0.02s user 0.02s system 97% cpu <b>0.038</b> total
</pre>



#### Implication

We do not print all the CRs anymore in the explainers. The question is if this was ever useful. If we want this kind of feature in the future we could extend the commands to not apply the CRs upon execution but rather only print them to stdout in case some flag is given. For the future **we should not do more than computing dumb strings within explainers**.